backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
PyJWT==2.9.0
requests==2.32.3
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
jinja2==3.1.4
